Organize the naming convention and minor code adjustment
